TDW-9 Added linting to developer packages, converted CSS to SCSS and and fixed linting errors
