+++ tbw-website/src/index.jsx
@@ -2,9 +2,6 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
 import "./i18n";
-import "./Styles/index.css";
+import "./Styles/index.scss";
 
-ReactDOM.render(
-  <App />,
-  document.getElementById("root")
-);
+ReactDOM.render(<App />, document.getElementById("root"));
